create common leagues state
